+++ src/imio/smartweb/core/tests/test_vocabularies.py
@@ -96,7 +96,7 @@ class TestVocabularies(ImioSmartwebTestCase):
         self.assertVocabularyLen("imio.smartweb.vocabulary.SubsiteDisplayMode", 3)
 
     def test_contact_blocks(self):
-        self.assertVocabularyLen("imio.smartweb.vocabulary.ContactBlocks", 9)
+        self.assertVocabularyLen("imio.smartweb.vocabulary.ContactBlocks", 10)
 
     @requests_mock.Mocker()
     def test_empty_remote_directory_entities(self, m):