+++ src/imio/smartweb/core/vocabularies.py
@@ -146,6 +146,7 @@ class ContactBlocksVocabularyFactory:
             ("titles", _("Title and Subtitle")),
             ("contact_informations", _("Contact informations")),
             ("address", _("Address")),
+            ("itinerary", _("Itinerary")),
             ("map", _("Map")),
             ("schedule", _("Schedule")),
             ("description", _("Description")),